window aligment feature added, introducing the new PixelPicker type
the type is used to define a single pixel of both windows (the
to-be-aligned and the align-according-to-that-one). the function itself
alignes the window to match the positions of the defined pixels.
it is super flexible, since it can calculate with the dimensions
of the windows and absolute pixel numbers at the same time. so you can
achieve any kind of alignment (side-by-side, vertical and/or horizontal
stacking, corner aligning, etc.)

i demo it with 3 crazy blinking windows in the corners and in the
middle bottom.