CursorWindow type and demo cursor added
the type itself is for a "singleton" variable named Cursor, which
gets the latest cursor positions automatically, and since it is
derived from Window, can be used to align other windows to it.

a crazy blinking demo cursor is added, which is a Window actually,
aligning right below the built-in Cursor window.